recommender/requirements.txt
fastapi==0.85.1
pyBKT==1.4
Pyrebase4==4.5.0
gunicorn==20.1.0
uvicorn==0.18.3
redis==4.3.4